lora-workflow/requirements.txt
fal-client>=0.4.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
replicate>=0.25.0
